delete flowbyseries.py, move series classes into flowby, flowbyactivity, flowbysector scripts
